refactor!: move files from root to app dir & change main folder name (now config)
