+++ app/blog/models.py
@@ -1,0 +1,12 @@
+from django.db import models
+
+
+class BlogModel(models.Model):
+    """Blog"""
+
+    title = models.CharField(max_length=150)
+    description = models.TextField(max_length=300)
+    date_creation = models.DateField()
+
+    def __str__(self):
+        return self.title
