+++ app/blog/templates/blog/all_blogs.html
@@ -1,0 +1,27 @@
+{% extends "my_portfolio/base.html" %}
+
+{% load static %}
+
+{% block content %}
+
+<h1 id="blogtitle" class="font-weight-bold text-center">Blog</h1>
+
+{% for blog in blogs %}
+<div class="row justify-content-center my-3">
+  <div class="col-md-10">
+    <a href="{% url 'blog:detail' blog.id %}">
+      <h3>
+          {{ blog.title }}
+      </h3>
+    </a>
+    <h6 class="text-muted">
+        {{ blog.date_creation|date:'M d Y'|upper }}
+    </h6>
+    <h5>
+        {{ blog.description|striptags|truncatechars:100 }}
+    </h5>
+  </div>
+</div>
+{% endfor %}
+
+{% endblock %}
