+++ app/common/views.py
@@ -1,0 +1,9 @@
+class TitleMixin:
+    """Common mixin for adding page title"""
+
+    title = None
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['title'] = self.title
+        return context
