+++ app/my_portfolio/admin.py
