--- portfolio/blog/templates/blog/detail.html
@@ -1,17 +1,0 @@
-{% extends "my_portfolio/base.html" %}
-
-{% load static %}
-
-{% block content %}
-
-<h1 class="text-center mt-5" id="blogdetailtitle">
-    {{ blog.title }}
-</h1>
-
-<h5 class="text-center text-muted mb-5">
-    {{ blog.date_creation|date:'F jS Y' }}
-</h5>
-
-{{ blog.description|safe }}
-
-{% endblock %}
