--- portfolio/common/views.py
@@ -1,7 +1,0 @@
-class TitleMixin:
-    title = None
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['title'] = self.title
-        return context
